监视重载的 tail 跟随一次性作废，失败不再留给下一份文档
问题：`monitor_pending`（监视页重载后该把光标/滚动归位到哪一页）只在成功臂被
取走。日志被轮转或删除导致重载失败时它一直挂着；之后关掉那一页、在同一个槽位
打开任意文件，那份新文档就会被陈旧意图强滚到文末、光标落到最后一行——用户只是
打开了一个文件，却发现自己站在文末。

原因：一次性意图的作废点写在了「兑现成功」那一条分支里，失败/丢弃出口忘了；
且它记的是 tab 下标，关页左移后比的是「现在的第 0 页」而不是当初那一页。

改法：与 `pending_link_goto` / `pending_fif_goto` 同批在 Loaded 入口摘走，兑现
与否取决于「本次装载真的落进了目标页」；字段改存页 id，归页判定时按 id 认领。

验证：全量绿、clippy 与 cargo doc 均 exit=0。新增
monitor_tail_follow_does_not_leak_into_the_next_load：开监视 → 巡检发起重载 →
重载失败 → 关页 → 同槽位打开另一个文件，断言光标仍在 (0,0) 且 scroll_top 未动；
把入口的 take() 改成只读（等价旧口径的不作废）即转红在「失败出口必须作废」。
既有 monitor_toggle_and_tick_reload_with_tail_follow（成功路径的 tail 跟随）与
monitor_skips_dirty_page 原样通过。

+++ crates/app/src/state.rs
@@ -236,11 +236,13 @@ pub(crate) struct Editpad {
     pub(crate) column_editor_visible: bool,
     /// 对话框草稿（跨开合保留）
     pub(crate) column_editor: ColumnEditorDraft,
-    /// P130：待归页的监视重载（tab 下标, 是否 tail 跟随, 重载前视图）——
+    /// P130：待归页的监视重载（**页 id**, 是否 tail 跟随, 重载前视图）——
     /// check_external_changes 发起监视重载时捕获，Loaded 归页时消费
-    ///（元组即重载前视图快照，一次性搬运不设类型别名）
+    ///（元组即重载前视图快照，一次性搬运不设类型别名）。
+    /// P221：曾是 tab 下标，且只在成功臂被取走——失败/丢弃出口留着它，
+    /// 下一次任意装载就会把「跟随到文末」用到无关文档上。
     #[allow(clippy::type_complexity)]
-    pub(crate) monitor_pending: Option<(usize, bool, Option<(usize, usize, f32, f32)>)>,
+    pub(crate) monitor_pending: Option<(u64, bool, Option<(usize, usize, f32, f32)>)>,
     /// P133：链接点击的待跳行号（1 起）——Ctrl+点击 `file.c:42` 时暂存，
     /// 下一次 Loaded 装载结算后一次性消费（jump_to_line）
     pub(crate) pending_link_goto: Option<u32>,
